requirements.txt
# Core
streamlit==1.39.0
pandas==2.2.3
duckdb==1.1.3
plotly==5.24.1

# AI/ML
openai==1.57.2
tiktoken==0.7.0
google-generativeai==0.7.2
duckduckgo-search==6.1.8

# Environment & HTTP
python-dotenv==1.0.1
requests==2.32.3

# Scheduling
APScheduler==3.11.0
pytz==2024.2

# Utils
beautifulsoup4==4.12.3
tabulate==0.9.0
google-search-results==2.4.2